move php extraction rules behind the indexer language registry
S3-PHP (ARCH-012), one language rollback unit, following S3-KOTLIN. PHP's
tree-sitter rule file, compiled-rule cache, member-call callsite marker,
namespace/type MEMBER collector, and receiver-call resolution engine move out
of the indexer god file into `codestory-indexer/src/languages/php.rs`, behind
the existing `LanguageExtraction` registry. Every language-keyed dispatch that
spelled "php" now reads the registry first and keeps its residual `match` for
the languages that have not moved; the last sibling package deletes those arms.

The move is mechanical: 28 functions leave `lib.rs` and land in
`languages/php.rs` byte-identical apart from three declared renames
(`collect_php_member_edges` -> `member_edge_specs`,
`collect_php_receiver_call_edges` -> `receiver_call_specs`, `php_member_call`
-> `member_call`) and the `pub(crate)` widening the first two need to stay
reachable. `split_case_insensitive_alias` and
`starts_with_case_insensitive_keyword` moved with them because PHP was their
only caller. `normalized_receiver_variable` and
`collect_enclosing_type_member_edges` stayed because other languages use them.

`lib.rs::language_member_specs` keeps its `"php"` arm because
`LanguageExtraction` has no manual-MEMBER field and adding one would rewrite
every sibling package's row; the arm now calls `languages::php::
member_edge_specs`, so the rule body moved even though the dispatch did not.
`lib.rs::receiver_annotation_required_callsite_marker` and
`resolution::is_php_member_call_placeholder` are repointed at
`languages::php::MEMBER_CALLSITE_MARKER` for the same reason.

Proof: `tests/language_extraction_snapshot.rs` gains a PHP row pinning the full
rendered projection of both PHP fixtures -- the fidelity_lab file through the
store pipeline (resolution included) and the tictactoe file through raw
`index_file`. Both goldens were captured before the production edit and are
byte-identical after it (sha256 eeb61741.. / d9388cf3..). Mutation-tested:
dropping `receiver_call_specs` reddens both snapshots, and changing
`qualified_name_delimiter` to "::" reddens 65 of 116 tictactoe lines while
`fidelity_regression` and `tictactoe_language_coverage` both stay green -- the
threshold suites alone would have waved that regression through.

Declared behavior changes:
- none -- output-equal move. PHP fidelity_regression and
  tictactoe_language_coverage projections are proven byte-identical before and
  after, and no packet_search_eval baseline is touched.
- `LanguageRuleset::Php`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, matching the Kotlin arm.
- `codestory_contracts::language_support::LANGUAGE_COMMENT_PROFILES` gains a
  `php` row and the CLI's residual roster drops its `"php"` arm. The rendered
  marker is `//` before and after, pinned by
  `comment_markers_stay_identical_after_the_registry_lookup`.
- No version surface touched; everything stays 0.16.3.

Closes #1687
