+++ crates/codestory-indexer/src/resolution/mod.rs
@@ -1441,7 +1441,7 @@ fn is_php_member_call_placeholder(edge_kind: EdgeKind, callsite_identity: Option
     callsite_identity.is_some_and(|identity| {
         identity
             .split('|')
-            .any(|part| part == crate::PHP_MEMBER_CALLSITE_MARKER)
+            .any(|part| part == crate::languages::php::MEMBER_CALLSITE_MARKER)
     })
 }
