+++ crates/codestory-indexer/tests/fixtures/language_snapshots/php_fidelity.txt
@@ -1,0 +1,39 @@
+node CLASS name=ConsoleNotifier qn=App.ConsoleNotifier canonical=<ROOT>/fidelity.php:App.ConsoleNotifier line=12 col=1 end=18:2 file=FILE:<ROOT>/fidelity.php@1
+node CLASS name=Event qn=App.Event canonical=<ROOT>/fidelity.php:App.Event line=28 col=1 end=33:2 file=FILE:<ROOT>/fidelity.php@1
+node CLASS name=Repository qn=App.Repository canonical=<ROOT>/fidelity.php:App.Repository line=20 col=1 end=26:2 file=FILE:<ROOT>/fidelity.php@1
+node CLASS name=Workflow qn=App.Workflow canonical=<ROOT>/fidelity.php:App.Workflow line=35 col=1 end=54:2 file=FILE:<ROOT>/fidelity.php@1
+node FILE name=<ROOT>/fidelity.php qn=<ROOT>/fidelity.php canonical=<ROOT>/fidelity.php:<ROOT>/fidelity.php:1 line=1 col=1 end=60:0 file=FILE:<ROOT>/fidelity.php@1
+node FUNCTION name=orchestrate_php qn=orchestrate_php canonical=<ROOT>/fidelity.php:orchestrate_php#0 line=56 col=1 end=60:2 file=FILE:<ROOT>/fidelity.php@1
+node INTERFACE name=Notifier qn=App.Notifier canonical=<ROOT>/fidelity.php:App.Notifier line=7 col=1 end=10:2 file=FILE:<ROOT>/fidelity.php@1
+node METHOD name=ConsoleNotifier.notify qn=App.ConsoleNotifier.notify canonical=<ROOT>/fidelity.php:App.ConsoleNotifier.notify#0 line=14 col=5 end=17:6 file=FILE:<ROOT>/fidelity.php@1
+node METHOD name=Event.__construct qn=App.Event.__construct canonical=<ROOT>/fidelity.php:App.Event.__construct#0 line=30 col=5 end=32:6 file=FILE:<ROOT>/fidelity.php@1
+node METHOD name=Notifier.notify qn=App.Notifier.notify canonical=<ROOT>/fidelity.php:App.Notifier.notify#0 line=9 col=5 end=9:48 file=FILE:<ROOT>/fidelity.php@1
+node METHOD name=Repository.save qn=App.Repository.save canonical=<ROOT>/fidelity.php:App.Repository.save#0 line=22 col=5 end=25:6 file=FILE:<ROOT>/fidelity.php@1
+node METHOD name=Workflow.__construct qn=App.Workflow.__construct canonical=<ROOT>/fidelity.php:App.Workflow.__construct#0 line=37 col=5 end=41:6 file=FILE:<ROOT>/fidelity.php@1
+node METHOD name=Workflow.decorate qn=App.Workflow.decorate canonical=<ROOT>/fidelity.php:App.Workflow.decorate#0 line=50 col=5 end=53:6 file=FILE:<ROOT>/fidelity.php@1
+node METHOD name=Workflow.run qn=App.Workflow.run canonical=<ROOT>/fidelity.php:App.Workflow.run#0 line=43 col=5 end=48:6 file=FILE:<ROOT>/fidelity.php@1
+node MODULE name=Random/Randomizer qn=Random/Randomizer canonical=<ROOT>/fidelity.php:Random/Randomizer#0 line=5 col=5 end=5:22 file=FILE:<ROOT>/fidelity.php@1
+node NAMESPACE name=App qn=App canonical=<ROOT>/fidelity.php:App#0 line=3 col=1 end=3:15 file=FILE:<ROOT>/fidelity.php@1
+node UNKNOWN name=decorate qn=decorate canonical=<ROOT>/fidelity.php:decorate#0 line=47 col=16 end=47:24 file=FILE:<ROOT>/fidelity.php@1
+node UNKNOWN name=getBytes qn=getBytes canonical=<ROOT>/fidelity.php:getBytes#0 line=59 col=50 end=59:58 file=FILE:<ROOT>/fidelity.php@1
+node UNKNOWN name=notify qn=notify canonical=<ROOT>/fidelity.php:notify#0 line=45 col=26 end=45:32 file=FILE:<ROOT>/fidelity.php@1
+node UNKNOWN name=run qn=run canonical=<ROOT>/fidelity.php:run#0 line=59 col=16 end=59:19 file=FILE:<ROOT>/fidelity.php@1
+node UNKNOWN name=save qn=save canonical=<ROOT>/fidelity.php:save#0 line=46 col=28 end=46:32 file=FILE:<ROOT>/fidelity.php@1
+edge CALL src=FUNCTION:orchestrate_php@56 dst=METHOD:Workflow.run@43 resolved_src=FUNCTION:orchestrate_php@56 resolved_dst=METHOD:Workflow.run@43 line=59 confidence=1.0000 certainty=Certain callsite=h0:59:1:h1 candidates=[]
+edge CALL src=FUNCTION:orchestrate_php@56 dst=UNKNOWN:getBytes@59 resolved_src=FUNCTION:orchestrate_php@56 resolved_dst=- line=59 confidence=- certainty=- callsite=h0:59:1:h2|syntax:php-member-call|receiver-owner:Randomizer|receiver-module:Random.Randomizer candidates=[]
+edge CALL src=METHOD:Workflow.run@43 dst=METHOD:Notifier.notify@9 resolved_src=METHOD:Workflow.run@43 resolved_dst=METHOD:Notifier.notify@9 line=45 confidence=1.0000 certainty=Certain callsite=h0:45:1:h3 candidates=[]
+edge CALL src=METHOD:Workflow.run@43 dst=METHOD:Repository.save@22 resolved_src=METHOD:Workflow.run@43 resolved_dst=METHOD:Repository.save@22 line=46 confidence=1.0000 certainty=Certain callsite=h0:46:1:h4 candidates=[]
+edge CALL src=METHOD:Workflow.run@43 dst=METHOD:Workflow.decorate@50 resolved_src=METHOD:Workflow.run@43 resolved_dst=METHOD:Workflow.decorate@50 line=47 confidence=1.0000 certainty=Certain callsite=h0:47:1:h5 candidates=[]
+edge IMPORT src=MODULE:Random/Randomizer@5 dst=MODULE:Random/Randomizer@5 resolved_src=MODULE:Random/Randomizer@5 resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:ConsoleNotifier@12 dst=METHOD:ConsoleNotifier.notify@14 resolved_src=- resolved_dst=- line=14 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Event@28 dst=METHOD:Event.__construct@30 resolved_src=- resolved_dst=- line=30 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Repository@20 dst=METHOD:Repository.save@22 resolved_src=- resolved_dst=- line=22 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@35 dst=METHOD:Workflow.__construct@37 resolved_src=- resolved_dst=- line=37 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@35 dst=METHOD:Workflow.decorate@50 resolved_src=- resolved_dst=- line=50 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@35 dst=METHOD:Workflow.run@43 resolved_src=- resolved_dst=- line=43 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=INTERFACE:Notifier@7 dst=METHOD:Notifier.notify@9 resolved_src=- resolved_dst=- line=9 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=NAMESPACE:App@3 dst=CLASS:ConsoleNotifier@12 resolved_src=- resolved_dst=- line=12 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=NAMESPACE:App@3 dst=CLASS:Event@28 resolved_src=- resolved_dst=- line=28 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=NAMESPACE:App@3 dst=CLASS:Repository@20 resolved_src=- resolved_dst=- line=20 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=NAMESPACE:App@3 dst=CLASS:Workflow@35 resolved_src=- resolved_dst=- line=35 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=NAMESPACE:App@3 dst=INTERFACE:Notifier@7 resolved_src=- resolved_dst=- line=7 confidence=- certainty=Certain callsite=- candidates=[]
